+++ backend/src/main/java/com/dubu/backend/plan/config/OdsayApiConfig.java
@@ -1,4 +1,4 @@
-package com.dubu.backend.plan.infra.client;
+package com.dubu.backend.plan.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
